+++ src/main/java/com/raoulvdberge/refinedstorage/api/util/IComparer.java
@@ -12,7 +12,6 @@ public interface IComparer {
     int COMPARE_DAMAGE = 1;
     int COMPARE_NBT = 2;
     int COMPARE_QUANTITY = 4;
-    int COMPARE_OREDICT = 8;
 
     /**
      * Compares two stacks by the given flags.
@@ -63,14 +62,5 @@ public interface IComparer {
      * @param right the right stack
      * @return true if the NBT tags of the two stacks are the same, false otherwise
      */
-    boolean isEqualNBT(@Nullable ItemStack left, @Nullable ItemStack right);
-
-    /**
-     * Compares two stacks and checks if they share the same ore dictionary entry.
-     *
-     * @param left  the left stack
-     * @param right the right stack
-     * @return true if the two stacks share the same ore dictionary entry
-     */
-    boolean isEqualOredict(@Nullable ItemStack left, @Nullable ItemStack right);
+    boolean isEqualNbt(@Nullable ItemStack left, @Nullable ItemStack right);
 }