+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/API.java
@@ -286,22 +286,22 @@ public class API implements IRSAPI {
     }
 
     @Override
-    public int getItemStackHashCode(ItemStack stack, boolean tag) {
+    public int getItemStackHashCode(ItemStack stack) {
         int result = stack.getItem().hashCode();
         result = 31 * result + (stack.getItemDamage() + 1);
 
-        if (tag && stack.hasTagCompound()) {
-            result = calcHashCode(stack.getTagCompound(), result);
+        if (stack.hasTagCompound()) {
+            result = getHashCode(stack.getTagCompound(), result);
         }
 
         return result;
     }
 
-    private int calcHashCode(NBTBase tag, int result) {
+    private int getHashCode(NBTBase tag, int result) {
         if (tag instanceof NBTTagCompound) {
-            result = calcHashCode((NBTTagCompound) tag, result);
+            result = getHashCode((NBTTagCompound) tag, result);
         } else if (tag instanceof NBTTagList) {
-            result = calcHashCode((NBTTagList) tag, result);
+            result = getHashCode((NBTTagList) tag, result);
         } else {
             result = 31 * result + tag.hashCode();
         }
@@ -309,18 +309,18 @@ public class API implements IRSAPI {
         return result;
     }
 
-    private int calcHashCode(NBTTagCompound tag, int result) {
+    private int getHashCode(NBTTagCompound tag, int result) {
         for (String key : tag.getKeySet()) {
             result = 31 * result + key.hashCode();
-            result = calcHashCode(tag.getTag(key), result);
+            result = getHashCode(tag.getTag(key), result);
         }
 
         return result;
     }
 
-    private int calcHashCode(NBTTagList tag, int result) {
+    private int getHashCode(NBTTagList tag, int result) {
         for (int i = 0; i < tag.tagCount(); ++i) {
-            result = calcHashCode(tag.get(i), result);
+            result = getHashCode(tag.get(i), result);
         }
 
         return result;
@@ -331,7 +331,7 @@ public class API implements IRSAPI {
         int result = stack.getFluid().hashCode();
 
         if (stack.tag != null) {
-            result = calcHashCode(stack.tag, result);
+            result = getHashCode(stack.tag, result);
         }
 
         return result;