+++ src/main/java/com/raoulvdberge/refinedstorage/gui/GuiDestructor.java
@@ -20,7 +20,6 @@ public class GuiDestructor extends GuiBase {
 
         addSideButton(new SideButtonCompare(this, TileDestructor.COMPARE, IComparer.COMPARE_DAMAGE));
         addSideButton(new SideButtonCompare(this, TileDestructor.COMPARE, IComparer.COMPARE_NBT));
-        addSideButton(new SideButtonCompare(this, TileDestructor.COMPARE, IComparer.COMPARE_OREDICT));
 
         addSideButton(new SideButtonDestructorPickup(this));
     }