+++ src/main/java/com/raoulvdberge/refinedstorage/gui/GuiExporter.java
@@ -20,7 +20,6 @@ public class GuiExporter extends GuiBase {
 
         addSideButton(new SideButtonCompare(this, TileExporter.COMPARE, IComparer.COMPARE_DAMAGE));
         addSideButton(new SideButtonCompare(this, TileExporter.COMPARE, IComparer.COMPARE_NBT));
-        addSideButton(new SideButtonCompare(this, TileExporter.COMPARE, IComparer.COMPARE_OREDICT));
     }
 
     @Override