+++ src/main/java/com/raoulvdberge/refinedstorage/gui/GuiImporter.java
@@ -23,7 +23,6 @@ public class GuiImporter extends GuiBase {
 
         addSideButton(new SideButtonCompare(this, TileImporter.COMPARE, IComparer.COMPARE_DAMAGE));
         addSideButton(new SideButtonCompare(this, TileImporter.COMPARE, IComparer.COMPARE_NBT));
-        addSideButton(new SideButtonCompare(this, TileImporter.COMPARE, IComparer.COMPARE_OREDICT));
     }
 
     @Override