+++ src/main/java/com/raoulvdberge/refinedstorage/gui/GuiInterface.java
@@ -17,7 +17,6 @@ public class GuiInterface extends GuiBase {
 
         addSideButton(new SideButtonCompare(this, TileInterface.COMPARE, IComparer.COMPARE_DAMAGE));
         addSideButton(new SideButtonCompare(this, TileInterface.COMPARE, IComparer.COMPARE_NBT));
-        addSideButton(new SideButtonCompare(this, TileInterface.COMPARE, IComparer.COMPARE_OREDICT));
     }
 
     @Override