+++ src/main/java/com/raoulvdberge/refinedstorage/gui/GuiStorageMonitor.java
@@ -17,7 +17,6 @@ public class GuiStorageMonitor extends GuiBase {
 
         addSideButton(new SideButtonCompare(this, TileStorageMonitor.COMPARE, IComparer.COMPARE_DAMAGE));
         addSideButton(new SideButtonCompare(this, TileStorageMonitor.COMPARE, IComparer.COMPARE_NBT));
-        addSideButton(new SideButtonCompare(this, TileStorageMonitor.COMPARE, IComparer.COMPARE_OREDICT));
     }
 
     @Override