+++ src/main/java/com/raoulvdberge/refinedstorage/gui/control/SideButtonCompare.java
@@ -38,16 +38,9 @@ public class SideButtonCompare extends SideButton {
             ty = 80;
         } else if (mask == IComparer.COMPARE_NBT) {
             ty = 48;
-        } else if (mask == IComparer.COMPARE_OREDICT) {
-            ty = 224;
         }
 
         int tx = (parameter.getValue() & mask) == mask ? 0 : 16;
-
-        // This is reversed in icons.png :D
-        if (mask == IComparer.COMPARE_OREDICT) {
-            tx = tx == 16 ? 0 : 16;
-        }
 
         gui.drawTexture(x, y, tx, ty, 16, 16);
     }